+++ mobile/library/cc/engine_builder.cc
@@ -327,7 +327,7 @@ EngineBuilder& EngineBuilder::addPlatformFilter(std::string name) {
   return *this;
 }
 
-EngineBuilder& EngineBuilder::addRuntimeGuard(std::string guard, bool value) {
+EngineBuilder& EngineBuilder::setRuntimeGuard(std::string guard, bool value) {
   runtime_guards_.push_back({guard, value});
   return *this;
 }